forge fmt

+++ script/external/IWHYPE9.sol
@@ -2,24 +2,24 @@
 pragma solidity ^0.8.20;
 
 interface IWHYPE9 {
-  event Approval(address indexed src, address indexed guy, uint wad);
-  event Transfer(address indexed src, address indexed dst, uint wad);
-  event Deposit(address indexed dst, uint wad);
-  event Withdrawal(address indexed src, uint wad);
+  event Approval(address indexed src, address indexed guy, uint256 wad);
+  event Transfer(address indexed src, address indexed dst, uint256 wad);
+  event Deposit(address indexed dst, uint256 wad);
+  event Withdrawal(address indexed src, uint256 wad);
 
   function deposit() external payable;
 
-  function withdraw(uint wad) external;
+  function withdraw(uint256 wad) external;
 
-  function totalSupply() external view returns (uint);
+  function totalSupply() external view returns (uint256);
 
-  function approve(address guy, uint wad) external returns (bool);
+  function approve(address guy, uint256 wad) external returns (bool);
 
-  function transfer(address dst, uint wad) external returns (bool);
+  function transfer(address dst, uint256 wad) external returns (bool);
 
-  function transferFrom(address src, address dst, uint wad) external returns (bool);
+  function transferFrom(address src, address dst, uint256 wad) external returns (bool);
 
-  function balanceOf(address guy) external view returns (uint);
+  function balanceOf(address guy) external view returns (uint256);
 
-  function allowance(address src, address guy) external view returns (uint);
+  function allowance(address src, address guy) external view returns (uint256);
 }
